+++ xcube/cli/common.py
